+++ FilieraAgricola/src/main/java/it/unicam/cs/filieraagricola/api/controller/elemento/ProdottoTrasformatoController.java
@@ -1,7 +1,10 @@
 package it.unicam.cs.filieraagricola.api.controller.elemento;
 
+import it.unicam.cs.filieraagricola.api.entities.elemento.Categoria;
 import it.unicam.cs.filieraagricola.api.entities.elemento.ProdottoTrasformato;
+import it.unicam.cs.filieraagricola.api.services.UserService;
 import it.unicam.cs.filieraagricola.api.services.elemento.ProdottoTrasformatoService;
+import it.unicam.cs.filieraagricola.api.services.gestore.CategoriaService;
 import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -9,31 +12,15 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/trasformatore")
+@RequestMapping("/operatore/trasformatore")
 public class ProdottoTrasformatoController {
 
     @Autowired
-    private final ProdottoTrasformatoService prodottoTrasformatoService;
-    public ProdottoTrasformatoController(ProdottoTrasformatoService prodottoTrasformatoService) {
-        this.prodottoTrasformatoService = prodottoTrasformatoService;
-    }
-
-/*     @PostConstruct
-    public void initSampleData() {
-       //Prodotto trasformato derivante da quello sopra
-        ProdottoBase prodottoBase = new ProdottoBase();
-        prodottoBase.setNome("Pummador");
-        prodottoBase.setCertificazioni("AA");
-        prodottoBase.setMetodiDiColtivazione("dsa");
-        prodottoRepository.save(prodottoBase);
-        ProdottoTrasformato passataPomodoro = new ProdottoTrasformato();
-        passataPomodoro.setNome("Passata Pomodoro");
-        passataPomodoro.setCertificazioni("DOP");
-        passataPomodoro.setProdottoBase(prodottoBase);
-        passataPomodoro.setProcessoTrasformazione("Schiacciato da pressa");
-        prodottoRepository.save(passataPomodoro);
-        System.out.println("Prodotto trasformato creato: " + passataPomodoro);
-    } */
+    private ProdottoTrasformatoService prodottoTrasformatoService;
+    @Autowired
+    private UserService userService;
+    @Autowired
+    private CategoriaService categoriaService;
 
     @GetMapping("/dashboard")
     public ResponseEntity<String> getDashboard() {
@@ -51,7 +38,11 @@ public class ProdottoTrasformatoController {
     @RequestMapping({"/prodottitrasformati/{id}"})
     public ResponseEntity<Object> getProduct(@PathVariable("id") int id) {
         if(prodottoTrasformatoService.getProdottoTrasformato(id) != null ){
-            return new ResponseEntity<>(prodottoTrasformatoService.getProdottoTrasformato(id), HttpStatus.FOUND);
+            if(prodottoTrasformatoService.getProdottoTrasformato(id).getOperatore() == userService.getCurrentUser()) {
+                return new ResponseEntity<>(prodottoTrasformatoService.getProdottoTrasformato(id), HttpStatus.FOUND);
+            } else {
+                return new ResponseEntity<>("Non sei autorizzato a vedere questo prodotto", HttpStatus.UNAUTHORIZED);
+            }
         }
         return new ResponseEntity<>("Prodotto trasformato non esistente", HttpStatus.NOT_FOUND);
     }
@@ -70,9 +61,14 @@ public class ProdottoTrasformatoController {
                                                       @RequestParam("certificazioni") String certificazioni,
                                                       @RequestParam("prodottoBase") int IDprodottoBase,
                                                       @RequestParam("descrizione") String descrizione,
-                                                      @RequestParam("prezzo") double prezzo)
+                                                      @RequestParam("prezzo") double prezzo,
+                                                      @RequestParam("categoria") String categoria)
     {
-        if(prodottoTrasformatoService.aggiungiProdottoTrasformato(nome, processoTrasformazione, certificazioni, IDprodottoBase, descrizione, prezzo)){
+        if(!categoriaService.existsSameCategoria(categoria)){
+            return new ResponseEntity<>("Categoria non esistente, Categorie esistenti: " + categoriaService.getAllCategorie(), HttpStatus.NOT_FOUND);
+        }
+        Categoria cat = categoriaService.getCategoriaByNome(categoria).get();
+        if(prodottoTrasformatoService.aggiungiProdottoTrasformato(nome, processoTrasformazione, certificazioni, IDprodottoBase, descrizione, prezzo, cat)){
             return new ResponseEntity<>("Prodotto creato", HttpStatus.CREATED);
         }
         return new ResponseEntity<>("Prodotto trasformato esistente", HttpStatus.CONFLICT);
@@ -80,6 +76,9 @@ public class ProdottoTrasformatoController {
 
     @RequestMapping({"/prodottitrasformati/elimina/{id}"})
     public ResponseEntity<Object> deleteProduct(@PathVariable("id") int id) {
+        if(prodottoTrasformatoService.getProdottoTrasformato(id).getOperatore() != userService.getCurrentUser()) {
+            return new ResponseEntity<>("Non sei autorizzato", HttpStatus.UNAUTHORIZED);
+        }
         prodottoTrasformatoService.deleteProdottoTrasformato(id);
         return new ResponseEntity<>("Prodotto trasformato eliminato", HttpStatus.OK);
     }
@@ -89,6 +88,9 @@ public class ProdottoTrasformatoController {
             method = {RequestMethod.PUT}
     )
     public ResponseEntity<Object> updateProduct(@RequestBody ProdottoTrasformato prodottoTrasformato) {
+        if(prodottoTrasformato.getOperatore() != userService.getCurrentUser()) {
+            return new ResponseEntity<>("Non sei autorizzato", HttpStatus.UNAUTHORIZED);
+        }
         if(prodottoTrasformatoService.aggiornaProdottoTrasformato(prodottoTrasformato)){
             return new ResponseEntity<>("Prodotto trasformato aggiornato", HttpStatus.FOUND);
         }
